requirements.txt
python-dotenv
telethon
requests
beautifulsoup4
twscrape